pyproject.toml
[build-system]
requires = ["uv_build>=0.9.15"]
build-backend = "uv_build"

[project]
name = "mysmtp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "envelope>=1.7.0",
    "matplotlib>=3.9.2",
    "pandas>=2.3.3",
    "psutil>=7.1.3",
    "pydantic<2",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "rocketry>=2.5.1",
    "tqdm>=4.67.1",
    "tyro>=1.0.0",
]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = [
"osx-64",
"linux-64",
]

[tool.pixi.pypi-dependencies]
mysmtp = { path = ".", editable = true }

[tool.pixi.tasks]
install = { cmd = "mkdir -p $HOME/.config/systemd/user && cp mysmtp.service $HOME/.config/systemd/user/mysmtp.service" }
enable = { cmd = "systemctl --user enable --now mysmtp.service" , depends-on = ["install"]}
status = { cmd = "systemctl --user status mysmtp" }
log = { cmd = "journalctl --user -u mysmtp -f" }
stop = { cmd = "systemctl --user stop mysmtp" }
